requirements.txt
certifi==2018.4.16
chardet==3.0.4
coverage==4.5.2
idna==2.6
requests>=2.20.0
responses==0.10.5
Unidecode==1.0.22
urllib3>=1.23
textdistance[extras]
pytz
python-dateutil==2.8.0